flask/requirements.txt
alembic==0.9.1
aniso8601==8.0.0
appdirs==1.4.3
attrs==20.3.0
click==6.7
dataclasses==0.8
Flask==1.1.2
Flask-Migrate==2.0.3
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
importlib-metadata==3.1.0
itsdangerous==0.24
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.0.6
MarkupSafe==1.0
packaging==16.8
psycopg2-binary==2.8.6
pyparsing==2.2.0
pyrsistent==0.17.3
python-dotenv==0.15.0
python-editor==1.0.3
pytz==2020.4
six
SQLAlchemy==1.1.9
SQLAlchemy-serializer==1.3.4.4
Werkzeug==1.0.1
zipp==3.4.0
